// api/generated/model/getMyFavoriteRecipesOrderBy.ts
/**
 * Generated by orval v8.12.3 🍺
 * Do not edit manually.
 * Cheffy API
 * OpenAPI spec version: 1.0.0
 */

export type GetMyFavoriteRecipesOrderBy = typeof GetMyFavoriteRecipesOrderBy[keyof typeof GetMyFavoriteRecipesOrderBy];


export const GetMyFavoriteRecipesOrderBy = {
  newest: 'newest',
  oldest: 'oldest',
} as const;